admin service file strutcture update
