+++ services/admin/services/byteToGigabyte.js
@@ -1,0 +1,5 @@
+const byteToGigabyte = (byteValue) => {
+    return parseFloat((byteValue * (9.31 * 10 ** -10)).toFixed(1));
+};
+
+module.exports = byteToGigabyte;
